fix uikit collision groups

+++ core/scene.c
@@ -596,7 +596,7 @@ CastResult scene_cast_result_default(void) {
 
 CastHitType scene_cast_ray(Scene *sc,
                            const Ray *worldRay,
-                           uint8_t groups,
+                           uint16_t groups,
                            const DoublyLinkedList *filterOutTransforms,
                            CastResult *result) {
 
@@ -731,7 +731,7 @@ CastHitType scene_cast_box(Scene *sc,
                            const Box *aabb,
                            const float3 *unit,
                            float maxDist,
-                           uint8_t groups,
+                           uint16_t groups,
                            const DoublyLinkedList *filterOutTransforms,
                            CastResult *result) {
